moved locales, added public to git ignore, standardized file names
