+++ src/components/index.js
@@ -3,7 +3,7 @@ export { default as Footer } from "./footer";
 export { default as Header } from "./header";
 export { default as Hero } from "./hero";
 export { default as i18n } from "./i18n";
-export { default as LanguageSwitcher } from "./languageSwitcher";
+export { default as LanguageSwitcher } from "./language-switcher";
 export { default as Layout } from "./layout";
 export { default as Link } from "./link";
 export { default as Section } from "./section";
